+++ src/commands/createStaticWebApp/ApiLocationStep.ts
@@ -7,14 +7,16 @@ import { AzureWizardPromptStep } from "vscode-azureextensionui";
 import { defaultApiName } from "../../constants";
 import { ext } from "../../extensionVariables";
 import { localize } from "../../utils/localize";
+import { addLocationTelemetry } from "./addLocationTelemetry";
 import { IStaticWebAppWizardContext } from "./IStaticWebAppWizardContext";
 
 export class ApiLocationStep extends AzureWizardPromptStep<IStaticWebAppWizardContext> {
     public async prompt(wizardContext: IStaticWebAppWizardContext): Promise<void> {
         wizardContext.apiLocation = (await ext.ui.showInputBox({
             value: defaultApiName,
-            prompt: localize('apiLocation', "Enter the location of your Azure Functions code"),
+            prompt: localize('apiLocation', "Enter the location of your Azure Functions code (Leave blank for no Azure Functions code)"),
         })).trim();
+        addLocationTelemetry(wizardContext, 'apiLocation', defaultApiName);
     }
 
     public shouldPrompt(wizardContext: IStaticWebAppWizardContext): boolean {