+++ src/tree/EnvironmentTreeItem.ts
@@ -3,11 +3,12 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { AzureParentTreeItem, IActionContext, TreeItemIconPath } from "vscode-azureextensionui";
+import { AzExtTreeItem, AzureParentTreeItem, IActionContext, TreeItemIconPath } from "vscode-azureextensionui";
 import { productionEnvironmentName } from "../constants";
 import { openUrl } from "../utils/openUrl";
 import { treeUtils } from "../utils/treeUtils";
 import { AppSettingsTreeItem } from "./AppSettingsTreeItem";
+import { AppSettingTreeItem } from "./AppSettingTreeItem";
 import { FunctionsTreeItem } from "./FunctionsTreeItem";
 import { IAzureResourceTreeItem } from "./IAzureResourceTreeItem";
 import { StaticWebAppTreeItem } from "./StaticWebAppTreeItem";
@@ -25,7 +26,6 @@ export type StaticEnvironment = {
 };
 
 export class EnvironmentTreeItem extends AzureParentTreeItem implements IAzureResourceTreeItem {
-
     public static contextValue: string = 'azureStaticEnvironment';
     public readonly contextValue: string = EnvironmentTreeItem.contextValue;
     public appSettingsTreeItem: AppSettingsTreeItem;
@@ -48,7 +48,7 @@ export class EnvironmentTreeItem extends AzureParentTreeItem implements IAzureRe
     }
 
     public get label(): string {
-        return this.data.properties.buildId === 'default' ? productionEnvironmentName : this.data.properties.pullRequestTitle;
+        return this.data.properties.buildId === 'default' ? productionEnvironmentName : `#${this.name} - ${this.data.properties.pullRequestTitle}`;
     }
 
     public get description(): string | undefined {
@@ -69,5 +69,18 @@ export class EnvironmentTreeItem extends AzureParentTreeItem implements IAzureRe
 
     public async browse(): Promise<void> {
         await openUrl(`https://${this.data.properties.hostname}`);
+    }
+
+    public async pickTreeItemImpl(expectedContextValues: (string | RegExp)[]): Promise<AzExtTreeItem | undefined> {
+        for (const expectedContextValue of expectedContextValues) {
+            switch (expectedContextValue) {
+                case AppSettingsTreeItem.contextValue:
+                case AppSettingTreeItem.contextValue:
+                    return this.appSettingsTreeItem;
+                default:
+            }
+        }
+
+        return undefined;
     }
 }