+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 
-## 0.1.0 - 2020-05-19
+## 0.1.0 - 2020-05-18
 
 ### Added
 - View, create, browse, and delete Azure Static Web Apps (Preview)
-- View Static Web Apps production and staging environments
-- Create APIs and endpoints for your local SPA project
+- View a static web app's environments
+- Create JavaScript/TypeScript HTTP Functions
+- View, add, edit, and delete Application Settings